Move background audio code out of form entry viewmodel
