+++ collect_app/src/test/java/org/odk/collect/android/formentry/FormEntryViewModelTest.java
@@ -1,7 +1,5 @@
 package org.odk.collect.android.formentry;
 
-import android.Manifest;
-
 import androidx.test.core.app.ApplicationProvider;
 
 import org.javarosa.core.model.FormIndex;
@@ -14,21 +12,14 @@ import org.odk.collect.android.exception.JavaRosaException;
 import org.odk.collect.android.formentry.audit.AuditEvent;
 import org.odk.collect.android.formentry.audit.AuditEventLogger;
 import org.odk.collect.android.javarosawrapper.FormController;
-import org.odk.collect.android.permissions.PermissionsChecker;
 import org.odk.collect.android.preferences.PreferencesProvider;
-import org.odk.collect.audiorecorder.recorder.Output;
-import org.odk.collect.audiorecorder.recording.AudioRecorder;
 import org.odk.collect.utilities.Clock;
 import org.robolectric.RobolectricTestRunner;
 
 import java.io.IOException;
-import java.util.HashSet;
-import java.util.function.BiConsumer;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.nullValue;
 import static org.mockito.Mockito.atMostOnce;
 import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
@@ -46,9 +37,6 @@ public class FormEntryViewModelTest {
     private FormIndex startingIndex;
     private AuditEventLogger auditEventLogger;
     private Clock clock;
-    private FakeRecordAudioActionRegistry recordAudioActionRegistry;
-    private AudioRecorder audioRecorder;
-    private PermissionsChecker permissionsChecker;
 
     @Before
     public void setup() {
@@ -62,11 +50,7 @@ public class FormEntryViewModelTest {
 
         clock = mock(Clock.class);
 
-        audioRecorder = mock(AudioRecorder.class);
-        recordAudioActionRegistry = new FakeRecordAudioActionRegistry();
-
-        permissionsChecker = mock(PermissionsChecker.class);
-        viewModel = new FormEntryViewModel(clock, mock(Analytics.class), new PreferencesProvider(ApplicationProvider.getApplicationContext()), audioRecorder, permissionsChecker, recordAudioActionRegistry);
+        viewModel = new FormEntryViewModel(clock, mock(Analytics.class), new PreferencesProvider(ApplicationProvider.getApplicationContext()));
         viewModel.formLoaded(formController);
     }
 
@@ -148,98 +132,5 @@ public class FormEntryViewModelTest {
         when(clock.getCurrentTime()).thenReturn(12345L);
         viewModel.openHierarchy();
         verify(auditEventLogger).logEvent(AuditEvent.AuditEventType.HIERARCHY, true, 12345L);
-    }
-
-    @Test
-    public void whenRecordAudioActionIsTriggered_whenQualityIsVoiceOnly_startsAMRRecording() {
-        when(permissionsChecker.isPermissionGranted(Manifest.permission.RECORD_AUDIO)).thenReturn(true);
-
-        TreeReference treeReference = new TreeReference();
-        recordAudioActionRegistry.listener.accept(treeReference, "voice-only");
-
-        verify(audioRecorder).start(new HashSet<TreeReference>() {
-            {
-                add(treeReference);
-            }
-        }, Output.AMR);
-    }
-
-    @Test
-    public void whenRecordAudioActionIsTriggered_whenQualityIsLow_startsAACLowRecording() {
-        when(permissionsChecker.isPermissionGranted(Manifest.permission.RECORD_AUDIO)).thenReturn(true);
-
-        TreeReference treeReference = new TreeReference();
-        recordAudioActionRegistry.listener.accept(treeReference, "low");
-
-        verify(audioRecorder).start(new HashSet<TreeReference>() {
-            {
-                add(treeReference);
-            }
-        }, Output.AAC_LOW);
-    }
-
-    @Test
-    public void whenRecordAudioActionIsTriggered_whenQualityIsMissings_startsAMRRecording() {
-        when(permissionsChecker.isPermissionGranted(Manifest.permission.RECORD_AUDIO)).thenReturn(true);
-
-        TreeReference treeReference = new TreeReference();
-        recordAudioActionRegistry.listener.accept(treeReference, null);
-
-        verify(audioRecorder).start(new HashSet<TreeReference>() {
-            {
-                add(treeReference);
-            }
-        }, Output.AMR);
-    }
-
-    @Test
-    public void grantAudioPermission_startsBackgroundRecording() {
-        when(permissionsChecker.isPermissionGranted(Manifest.permission.RECORD_AUDIO)).thenReturn(false);
-
-        TreeReference treeReference1 = new TreeReference();
-        TreeReference treeReference2 = new TreeReference();
-        recordAudioActionRegistry.listener.accept(treeReference1, "low");
-        recordAudioActionRegistry.listener.accept(treeReference2, "low");
-
-        viewModel.grantAudioPermission();
-        verify(audioRecorder).start(new HashSet<TreeReference>() {
-            {
-                add(treeReference1);
-                add(treeReference2);
-            }
-        }, Output.AAC_LOW);
-    }
-
-    @Test
-    public void grantAudioPermission_clearsErrror() {
-        when(permissionsChecker.isPermissionGranted(Manifest.permission.RECORD_AUDIO)).thenReturn(false);
-
-        TreeReference treeReference1 = new TreeReference();
-        recordAudioActionRegistry.listener.accept(treeReference1, "low");
-
-        viewModel.grantAudioPermission();
-        assertThat(viewModel.getError().getValue(), is(nullValue()));
-    }
-
-    @Test
-    public void onCleared_unregistersRecordAudioActionListener() {
-        viewModel.onCleared();
-        assertThat(recordAudioActionRegistry.listener, is(nullValue()));
-    }
-
-    private static class FakeRecordAudioActionRegistry implements FormEntryViewModel.RecordAudioActionRegistry {
-
-
-        private BiConsumer<TreeReference, String> listener;
-
-        @Override
-        public void register(BiConsumer<TreeReference, String> listener) {
-            this.listener = listener;
-        }
-
-        @Override
-        public void unregister() {
-            this.listener = null;
-        }
     }
 }
